requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

torch==2.5.1
accelerate==1.9.0
bayesian-torch==0.5.0
pyyaml==6.0.1

# Scientific computing
numpy==1.26.3
pandas==2.2.0
h5py==3.10.0
netCDF4==1.6.5
vtk==9.4.2

# Eval and vis
nerfacc==0.5.3
lpips==0.1.4
pytorch-msssim==1.0.0
Pillow==11.1.0
tensorboard==2.15.1
matplotlib==3.8.2